ctr_lrw: Slow down queue processing
Scenario when Murata modem was reportin ERR=-18 (busy) has been
observed. Probably, this could be due to fast AT command feeding to
the modem. This fix puts a 3-second delay on each command processing
in the queues. This might be a temporary solution but we better
observe behavior in a safe manner.

Signed-off-by: Pavel Hübner <pavel.hubner@hardwario.com>