subsys: ctr_config: Fix flash erase sequence
The whole implementation is greatly simplified. The current
implementation was triggering Hard Fault exception after
the NCS v2.0.0 update.

There are two methods in the implementation:
1. Flash area erase (the whole storage partition).
2. NVS mount + clear.

I have tested both and they do the same thing. For now, we stick with
the simpler one. We will delete the other one in the future unless
we need to get back to it for some reason.

Signed-off-by: Pavel Hübner <pavel.hubner@hardwario.com>