subsys: ctr_ble: Fix max ID count
Previously, the number of possible connections / pairs did not have
enough allocated buffers. A memory overflow was likely to occur.

The commit also simplifies Kconfig options and improves the order
of execution in the BLE subsystem.

This is a major fix and all users are advised to use this update as soon
possible.

Signed-off-by: Pavel Hübner <pavel.hubner@hardwario.com>