drivers: chester_k: Add ADC Kconfig constraints
If ctr_batt was enabled altogether with ctr_k,
it was a conflicting settings. That was because ctr_batt enables
CONFIG_ADC, which implicitly enables SoC ADC driver. But ctr_k
requires a manual ADC control and there was an interfering
initialization.

On the other hand, ctr_batt only needs ADC API as it uses I2C
ADC driver,  but does not need SoC peripheral. Hereby, we should
explicitly disable the SoC ADC driver and related options.

Signed-off-by: Pavel Hübner <pavel.hubner@hardwario.com>